Put the Accent enum into a different class.
The Qt designer can only detect enum properties if the enum is part of a class that inherits QObject. As OfficeWidget can not inherit QObject for obvious reasons, we need to put it into a dummy class that inherits from QObject. The new class "Office" will now contain every enum within QOffice.